refactor: calendar to DatePicker -Add calendar primitive -Refactor calendar to DatePicker -Remove unused atributes from SelectDate
