+++ frontend/src/components/Primitives/Inputs/SelectDate/SelectDate.tsx
@@ -14,30 +14,21 @@ export type SearchInputProps = {
   placeholder: string;
   disabled?: boolean;
   currentValue?: string;
-  handleChange?: (event: React.ChangeEvent<HTMLInputElement>) => void;
-  handleClear?: () => void;
 };
 
-const SelectDate = ({
-  placeholder,
-  disabled = false,
-  currentValue,
-  handleChange,
-  handleClear,
-}: SearchInputProps) => (
+const SelectDate = ({ placeholder, disabled = false, currentValue }: SearchInputProps) => (
   <Flex
     css={{ position: 'relative', width: '100%', height: 'auto' }}
     direction="column"
-    data-testid="searchInput"
+    data-testid="selectDate"
   >
     <StyledInputWrapper gap="8" disabled={disabled} style={{ backgroundColor: 'white' }}>
       <Flex css={{ flexGrow: '1', backgroundColor: 'white' }}>
         <StyledInput
-          id="searchInput"
+          id="selectDate"
           disabled={false}
           placeholder=" "
           value={currentValue}
-          onChange={handleChange}
           css={{ pt: '$28', pb: '$8' }}
           style={{ cursor: 'pointer' }}
         />
@@ -45,25 +36,13 @@ const SelectDate = ({
           {placeholder}
         </PlaceholderText>
       </Flex>
-      <IconWrapper
-        justify="center"
-        data-type="clear"
-        css={{
-          width: '$24',
-          height: '$24',
-          borderRadius: '100%',
-        }}
-        onClick={handleClear}
-        data-testid="searchInputClear"
-      >
+      <IconWrapper>
         <Icon
           name="calendar-alt"
           size={24}
           css={{
-            color: '$grey',
+            color: '$primary300',
             backgroundColor: 'transparent',
-            borderRadius: '100%',
-            padding: '$2',
           }}
         />
       </IconWrapper>
